--- plugins/source/aws/resources/services/ecr/registries_fetch.go
@@ -1,20 +1,0 @@
-package ecr
-
-import (
-	"context"
-
-	"github.com/aws/aws-sdk-go-v2/service/ecr"
-	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
-	"github.com/cloudquery/plugin-sdk/schema"
-)
-
-func fetchEcrRegistries(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
-	c := meta.(*client.Client)
-	svc := c.Services().Ecr
-	output, err := svc.DescribeRegistry(ctx, &ecr.DescribeRegistryInput{})
-	if err != nil {
-		return err
-	}
-	res <- output
-	return nil
-}
