+++ plugins/source/aws/resources/services/ecr/repositories.go
@@ -1,6 +1,10 @@
 package ecr
 
 import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/ecr"
 	"github.com/aws/aws-sdk-go-v2/service/ecr/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
@@ -42,4 +46,45 @@ func Repositories() *schema.Table {
 			RepositoryImages(),
 		},
 	}
+}
+func fetchEcrRepositories(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	svc := meta.(*client.Client).Services().Ecr
+	paginator := ecr.NewDescribeRepositoriesPaginator(svc, &ecr.DescribeRepositoriesInput{
+		MaxResults: aws.Int32(1000),
+	})
+	for paginator.HasMorePages() {
+		page, err := paginator.NextPage(ctx)
+
+		if err != nil {
+			return err
+		}
+		res <- page.Repositories
+	}
+	return nil
+}
+
+func resolveRepositoryTags(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
+	output, err := meta.(*client.Client).Services().Ecr.ListTagsForResource(ctx, &ecr.ListTagsForResourceInput{
+		ResourceArn: resource.Item.(types.Repository).RepositoryArn,
+	})
+	if err != nil {
+		return err
+	}
+	return resource.Set(c.Name, client.TagsToMap(output.Tags))
+}
+
+func resolveRepositoryPolicy(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
+	svc := meta.(*client.Client).Services().Ecr
+	repo := resource.Item.(types.Repository)
+	output, err := svc.GetRepositoryPolicy(ctx, &ecr.GetRepositoryPolicyInput{
+		RepositoryName: repo.RepositoryName,
+		RegistryId:     repo.RegistryId,
+	})
+	if err != nil {
+		if client.IsAWSError(err, "RepositoryPolicyNotFoundException") {
+			return nil
+		}
+		return err
+	}
+	return resource.Set(c.Name, output.PolicyText)
 }
