+++ plugins/source/aws/resources/services/ecr/repository_image_scan_findings.go
@@ -1,6 +1,11 @@
 package ecr
 
 import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/ecr"
+	"github.com/aws/aws-sdk-go-v2/service/ecr/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/resources/services/ecr/models"
 	"github.com/cloudquery/plugin-sdk/schema"
@@ -18,4 +23,39 @@ func RepositoryImageScanFindings() *schema.Table {
 			client.DefaultRegionColumn(false),
 		},
 	}
+}
+func fetchEcrRepositoryImageScanFindings(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	image := parent.Item.(types.ImageDetail)
+	repo := parent.Parent.Item.(types.Repository)
+	for _, tag := range image.ImageTags {
+		config := ecr.DescribeImageScanFindingsInput{
+			RepositoryName: repo.RepositoryName,
+			ImageId: &types.ImageIdentifier{
+				ImageDigest: image.ImageDigest,
+				ImageTag:    aws.String(tag),
+			},
+			MaxResults: aws.Int32(1000),
+		}
+
+		paginator := ecr.NewDescribeImageScanFindingsPaginator(meta.(*client.Client).Services().Ecr, &config)
+		for paginator.HasMorePages() {
+			output, err := paginator.NextPage(ctx)
+			if err != nil {
+				if client.IsAWSError(err, "ScanNotFoundException") {
+					return nil
+				}
+				return err
+			}
+			res <- models.ImageScanWrapper{
+				ImageScanFindings: output.ImageScanFindings,
+				ImageTag:          aws.String(tag),
+				ImageDigest:       image.ImageDigest,
+				ImageScanStatus:   output.ImageScanStatus,
+				RegistryId:        repo.RegistryId,
+				RepositoryName:    repo.RepositoryName,
+			}
+		}
+	}
+
+	return nil
 }
